+++ Sdl.h
@@ -4,7 +4,7 @@
 
 Sdl setup(const int res, const int fps, const char* const name);
 void release(const Sdl sdl);
-Sdl tick(const Sdl sdl, const unsigned long long renders);
+Sdl tick(const Sdl sdl, const int renders);
 void churn(const Sdl sdl);
 void present(const Sdl sdl);
 void render(const Sdl sdl, const Hero hero, const Sprites sprites, const Map map);
